pyproject.toml
[project]
name = "explore-anthropic-sdk"
version = "0.1.0"
description = "Small scripts for learning the Anthropic Python SDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.97.0",
    "python-dotenv>=1.2.2",
]
